feat: finish modal flow
