+++ src/components/Modal/styles.ts
@@ -1,0 +1,3 @@
+import styled from 'styled-components/native';
+
+export const Container = styled.Modal``;
